[#258] Add home screen
