+++ bricks/template/__brick__/{{project_name.snakeCase()}}/lib/app/screens/home/home_screen.dart
@@ -1,0 +1,75 @@
+import 'package:{{project_name.snakeCase()}}/app/screens/home/home_view_model.dart';
+import 'package:{{project_name.snakeCase()}}/app/screens/home/home_view_state.dart';
+import 'package:{{project_name.snakeCase()}}/domain/usecases/get_users_use_case.dart';
+import 'package:{{project_name.snakeCase()}}/main.dart';
+import 'package:flutter_config/flutter_config.dart';
+
+final homeViewModelProvider =
+StateNotifierProvider.autoDispose<HomeViewModel, HomeViewState>((ref) {
+  return HomeViewModel(
+    getIt.get<GetUsersUseCase>(),
+  );
+});
+
+class HomeScreen extends ConsumerStatefulWidget {
+  const HomeScreen({super.key});
+
+  @override
+  _HomeScreenState createState() => _HomeScreenState();
+}
+
+class _HomeScreenState extends ConsumerState<HomeScreen> {
+  @override
+  void initState() {
+    super.initState();
+    ref.read(homeViewModelProvider.notifier).getUsers();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: FutureBuilder<PackageInfo>(
+            future: PackageInfo.fromPlatform(),
+            builder: (context, snapshot) {
+              return snapshot.hasData
+                  ? Text(snapshot.data?.appName ?? "")
+                  : const SizedBox.shrink();
+            }),
+      ),
+      body: Center(
+        child: Column(
+          children: [
+            const SizedBox(height: 24),
+            FractionallySizedBox(
+              widthFactor: 0.5,
+              child: Image.asset(
+                Assets.images.nimbleLogo.path,
+                fit: BoxFit.fitWidth,
+              ),
+            ),
+            const SizedBox(height: 8),
+            Assets.svg.flutterLogo.svg(
+              width: 32,
+              height: 32,
+            ),
+            const SizedBox(height: 24),
+            Text(AppLocalizations.of(context)!.hello),
+            Text(
+              FlutterConfig.get('SECRET'),
+              style: const TextStyle(
+                color: AppColors.nimblePrimaryBlue,
+                fontSize: 24,
+              ),
+            ),
+            const SizedBox(height: 24),
+            ElevatedButton(
+              onPressed: () => context.go('/$routePathSecondScreen'),
+              child: const Text("Navigate to Second Screen"),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
