+++ bricks/template/__brick__/{{project_name.snakeCase()}}/lib/domain/usecases/get_users_use_case.dart
