+++ bricks/template/__brick__/{{project_name.snakeCase()}}/test/viewmodel/home_view_model_test.dart
@@ -3,7 +3,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:{{project_name.snakeCase()}}/app/screens/home/home_view_model.dart';
 import 'package:{{project_name.snakeCase()}}/domain/usecases/base/base_use_case.dart';
-import 'package:{{project_name.snakeCase()}}/main.dart';
+import 'package:{{project_name.snakeCase()}}/app/screens/home/home_screen.dart';
 
 import '../mocks/generate_mocks.mocks.dart';
 import '../mocks/response/user_response_mocks.dart';
